Use int instead of uint for config params
  * See: https://git.sr.ht/~jamesponddotco/imgdiet-go/commit/7b6744823c0b0aa17189833e4b9b6469c79665af

+++ config.go
@@ -15,7 +15,7 @@ type Config struct {
 	LogLevel vips.LogLevel
 
 	// Cache defines the size of the libvips cache in bytes.
-	Cache uint64
+	Cache int
 
 	// MaxConcurrency defines the maximum number of concurrent operations that
 	// libvips can perform.
